requirements.txt
scikit-learn==0.24.2
matplotlib==3.3.3
psycopg
pandas
datetime
requests
pandasql
urllib3
traceback2